Fix: Resolve build errors after frontend refactoring
This commit addresses several critical TypeScript and import path errors that arose following the major refactoring of the frontend module structures (Procurement, Assets, Service Requests) and theming.

Corrections include:
1.  **Theme Error:**
    *   Resolved "used before declaration" error for `commonDarkModeComponents` and `commonLightModeComponents` in `itsm_frontend/src/theme/theme.ts` by reordering definitions.

2.  **Import Path Corrections:**
    *   Systematically adjusted relative import paths (e.g., `../../../` to `../../../../` or `../../../` as appropriate) in components within the Procurement, Assets, and Service Requests modules to correctly point to global resources (context, api, utils) from their new, potentially deeper, file locations.

3.  **Type Export/Import Verification:**
    *   Ensured that types moved to module-specific files (e.g., `PurchaseRequestDecisionData` in `procurementTypes.ts`) are correctly exported and imported in components, resolving issues where types were sought from old API file locations.

4.  **Implicit `any` Type Resolution:**
    *   Addressed "Parameter 'item' implicitly has an 'any' type" errors in `PurchaseOrderForm.tsx` by providing explicit `OrderItemData` type annotations in `map` and `reduce` callbacks.

These fixes aim to restore the build integrity of the frontend application and allow further development.
